display the comments on details click
