+++ src/app/modules/utilities.js
@@ -31,3 +31,17 @@ export const error = (message = '') => {
   const error = `An error has ocurred ${message}`;
   console.log(error);
 };
+
+export const getComments = async (beerId) => {
+  try {
+    const response = await apiCall(`${appID}/comments?item_id=${beerId}`, 'GET', {}, true);
+    const res = JSON.parse(response);
+    if (res.error) {
+      throw new Error(res.error);
+    }
+    return JSON.parse(response);
+  } catch (err) {
+    error(err);
+    return [];
+  }
+};
